# Conflicts:
#   backend/main-service/src/main/java/com/turtlecoin/mainservice/MainServiceApplication.java
#   backend/main-service/src/main/java/com/turtlecoin/mainservice/domain/transaction/repository/TransactionRepository.java
#   backend/main-service/src/main/java/com/turtlecoin/mainservice/domain/transaction/service/TransactionService.java

+++ backend/main-service/src/main/java/com/turtlecoin/mainservice/domain/turtle/dto/AuctionTurtleInfoDTO.java
@@ -10,5 +10,5 @@ import lombok.*;
 public class AuctionTurtleInfoDTO {
     private Long id;
     private Gender gender;
-    private Double size;
+    private int size;
 }
